if user gets delete we need to remove data in club_permissions and also api keys which were created by this user

+++ application/models/User_model.php
@@ -412,7 +412,8 @@ class User_Model extends CI_Model {
 			}
 			// Delete QSOs from $this->config->item('table_name')
 			$this->db->query("DELETE FROM bandxuser WHERE userid = ?",$user_id);
-			$this->db->query("DELETE FROM api WHERE user_id = ?",$user_id);
+			$this->db->query("DELETE FROM api WHERE user_id = ? OR created_by = ?", [$user_id, $user_id]);
+			$this->db->query("DELETE FROM club_permissions WHERE user_id = ? OR club_id = ?", [$user_id, $user_id]);
 			$this->db->query("DELETE FROM cat WHERE user_id = ?",$user_id);
 			$this->db->query("DELETE FROM lotw_certs WHERE user_id = ?",$user_id);
 			$this->db->query("DELETE FROM notes WHERE user_id = ?",$user_id);
@@ -421,7 +422,7 @@ class User_Model extends CI_Model {
 			$this->db->query("DELETE FROM queries WHERE userid = ?",$user_id);
 			$this->db->query("DELETE FROM station_profile WHERE user_id = ?",$user_id);
 			$this->db->query("DELETE FROM station_logbooks WHERE user_id = ?",$user_id);
-			$this->db->query("delete from user_options where user_id=?",$user_id);
+			$this->db->query("DELETE FROM user_options WHERE user_id=?",$user_id);
 			$this->db->query("DELETE FROM ".$this->config->item('auth_table')." WHERE user_id = ?",$user_id);
 			return 1;
 		} else {
